created utils_outliers; added interpolation of remaining outliers; combined segment length and abs value to determine outliers
