+++ df3dPostProcessing/df3dPostProcessing.py
@@ -1,5 +1,3 @@
-import itertools
-
 import numpy as np
 import os
 import pickle
@@ -8,8 +6,8 @@ from pathlib import Path
 from .utils.utils_alignment import align_3d
 from .utils.utils_angles import calculate_angles
 from .utils.utils_ball_info import ball_size_and_pos
+from .utils.utils_outliers import correct_outliers
 #from .utils.utils_plots import *
-import deepfly.signal_util
 
 df3d_skeleton = ['RFCoxa',
                  'RFFemur',
@@ -206,7 +204,7 @@ zero_pose_nmf = {'RF_leg':{'ThC_yaw':0,
                            'TiTa_pitch':-np.pi}}
 
 class df3dPostProcess:
-    def __init__(self, exp_dir, multiple = False, file_name = '', skeleton='df3d', calculate_3d=False):
+    def __init__(self, exp_dir, multiple = False, file_name = '', skeleton='df3d', calculate_3d=False, outlier_correction=False):
         self.exp_dir = exp_dir
         self.raw_data_3d = np.array([])
         self.raw_data_2d = np.array([])
@@ -214,7 +212,7 @@ class df3dPostProcess:
         self.template = template_nmf
         self.zero_pose = zero_pose_nmf
         self.raw_data_cams = {}
-        self.load_data(exp_dir, calculate_3d, skeleton, multiple, file_name)
+        self.load_data(exp_dir, calculate_3d, skeleton, multiple, file_name, outlier_correction=outlier_correction)
         self.data_3d_dict = load_data_to_dict(self.raw_data_3d, skeleton)
         self.data_2d_dict = load_data_to_dict(self.raw_data_2d, skeleton)
         self.aligned_model = {}
@@ -289,7 +287,7 @@ class df3dPostProcess:
 
         return ball_info
     
-    def load_data(self, exp, calculate_3d, skeleton, multiple, file_name):
+    def load_data(self, exp, calculate_3d, skeleton, multiple, file_name, outlier_correction):
         if multiple:
             currentDirectory = os.getcwd()
             dataFile = os.path.join(currentDirectory,file_name)
@@ -300,7 +298,7 @@ class df3dPostProcess:
             if skeleton == 'prism' or skeleton == 'lp3d':
                 data={'points3d_wo_procrustes':data}
             if calculate_3d:
-                self.raw_data_3d = triangulate_2d(data, exp)
+                self.raw_data_3d = triangulate_2d(data, exp, outlier_correction)
                 
             for key, vals in data.items():
                 if not isinstance(key,str):
@@ -311,7 +309,7 @@ class df3dPostProcess:
                 elif key == 'points2d':
                     self.raw_data_2d = vals
 
-def triangulate_2d(data, exp_dir):
+def triangulate_2d(data, exp_dir, outlier_correction):
     img_folder = exp_dir[:exp_dir.find('df3d')]
     out_folder = exp_dir[:exp_dir.find('pose_result')]
     num_images = data['points3d'].shape[0]
@@ -325,129 +323,11 @@ def triangulate_2d(data, exp_dir):
         camNet.cam_list[cam_id].set_tvec(data[cam_id]["tvec"]) 
     camNet.triangulate() 
 
-    camNet = correct_outliers(camNet)
+    if outlier_correction:
+        print("correcting")
+        camNet = correct_outliers(camNet)
 
     return camNet.points3d
-
-def correct_outliers(camNet):
-    from deepfly.cv_util import triangulate_linear
-
-    #outlier_image_ids = np.where(np.abs(camNet.points3d) > 5)[0]
-    #outlier_joint_ids = np.where(np.abs(camNet.points3d) > 5)[1]
-    #print(outlier_image_ids)
-    #print(outlier_joint_ids)
-
-    start_indices = np.array([0, 1, 2, 3,
-                              5, 6, 7, 8,
-                              10, 11, 12, 13,
-                              19, 20, 21, 22,
-                              24, 25, 26, 27,
-                              29, 30, 31, 32,],
-                             dtype=np.int)
-    stop_indices = start_indices + 1
-    #pairs = np.vstack((start_indices, stop_indices))
-    #lengths = np.sum((camNet.points3d[:, start_indices, :] - camNet.points3d[:, stop_indices, :]) ** 2, axis=2)
-    lengths = np.linalg.norm(camNet.points3d[:, start_indices, :] - camNet.points3d[:, stop_indices, :], axis=2)
-    median_lengths = np.median(lengths, axis=0)
-    #length_outliers = (lengths > np.quantile(lengths, 0.99, axis=0)) | (lengths < np.quantile(lengths, 0.01, axis=0))
-    length_outliers = (lengths > median_lengths * 1.4) | (lengths < median_lengths * 0.4)
-    #print(np.quantile(lengths, 0.99, axis=0))
-    #print(np.max(lengths, axis=0))
-    #print(np.quantile(lengths, 0.01, axis=0))
-    #print(np.min(lengths, axis=0))
-    outlier_mask = np.zeros(camNet.points3d.shape, dtype=np.bool)
-    for i, mask_offset in enumerate([0, 5, 10, 19, 24, 29]):
-        claw_outliers = np.where(length_outliers[:, i * 4 + 4] & ~length_outliers[:, i * 4 + 3])[0]
-        tarsus_outliers = np.where(length_outliers[:, i * 4 + 3] & length_outliers[:, i * 4 + 2])[0]
-        tibia_outliers = np.where(length_outliers[:, i * 4 + 2] & length_outliers[:, i * 4 + 1])[0]
-        femur_outliers = np.where(length_outliers[:, i * 4 + 1] & length_outliers[:, i * 4 + 0])[0]
-        outlier_mask[femur_outliers, 1 + mask_offset] = True
-        outlier_mask[tibia_outliers, 2 + mask_offset] = True
-        outlier_mask[tarsus_outliers, 3 + mask_offset] = True
-        outlier_mask[claw_outliers, 4 + mask_offset] = True
-    
-    outlier_image_ids = np.where(outlier_mask)[0]
-    outlier_joint_ids = np.where(outlier_mask)[1]
-    #print(outlier_image_ids)
-    #print(outlier_joint_ids)
-    #exit()
-    
-
-    def _triangluate_specific_cameras(camNet, cam_id_list,img_id, j_id):
-        cam_list_iter = list()
-        points2d_iter = list()
-        for cam in [camNet.cam_list[cam_idx] for cam_idx in cam_id_list]:
-            cam_list_iter.append(cam)
-            points2d_iter.append(cam[img_id, j_id, :])
-        return triangulate_linear(cam_list_iter, points2d_iter)
-
-    for img_id, joint_id in zip(outlier_image_ids, outlier_joint_ids):
-        reprojection_errors = list()
-        segment_length_diff = list()
-        points_using_2_cams = list()
-        # Select cameras based on which side the joint is on
-        all_cam_ids = [0, 1, 2] if joint_id < 19 else [4, 5, 6]
-        for subset_cam_ids in itertools.combinations(all_cam_ids, 2):
-            points3d_using_2_cams = _triangluate_specific_cameras(camNet, subset_cam_ids, img_id, joint_id)
-
-            new_diff = 0
-            median_index = np.where(stop_indices == joint_id)[0]
-            if len(median_index) > 0:
-                new_diff += np.linalg.norm(points3d_using_2_cams - camNet.points3d[img_id, joint_id - 1]) - median_lengths[median_index]
-            median_index = np.where(start_indices == joint_id)[0]
-            if len(median_index) > 0:
-                new_diff += np.linalg.norm(points3d_using_2_cams - camNet.points3d[img_id, joint_id + 1]) - median_lengths[median_index]
-            segment_length_diff.append(new_diff)
-
-            reprojection_error_function = lambda cam_id: camNet.cam_list[cam_id].project(points3d_using_2_cams) - camNet.cam_list[cam_id].points2d[img_id, joint_id]
-            reprojection_error = np.mean([reprojection_error_function(cam_id) for cam_id in subset_cam_ids])
-            reprojection_errors.append(reprojection_error)
-            points_using_2_cams.append(points3d_using_2_cams)
-
-        #reprojection_error_function = lambda cam_id: camNet.cam_list[cam_id].project(camNet.points3d[img_id, joint_id]) - camNet.cam_list[cam_id].points2d[img_id, joint_id]
-        #reprojection_error_3_cams = np.mean([reprojection_error_function(cam_id) for cam_id in all_cam_ids])
-
-        #if np.min(reprojection_errors) < reprojection_error_3_cams:
-        # Replace 3D points with best estimation from 2 cameras only
-        #best_cam_tuple_index = np.argmin(reprojection_errors)
-        best_cam_tuple_index = np.argmin(segment_length_diff)
-
-        old_diff = 0
-        new_diff = 0
-        median_index = np.where(stop_indices == joint_id)[0]
-        if len(median_index) > 0:
-            #print(pairs[:, median_index], joint_id, joint_id - 1)
-            old_diff += np.linalg.norm(camNet.points3d[img_id, joint_id] - camNet.points3d[img_id, joint_id - 1]) - median_lengths[median_index]
-            new_diff += np.linalg.norm(points_using_2_cams[best_cam_tuple_index] - camNet.points3d[img_id, joint_id - 1]) - median_lengths[median_index]
-        median_index = np.where(start_indices == joint_id)[0]
-        if len(median_index) > 0:
-            #print(pairs[:, median_index], joint_id, joint_id + 1)
-            old_diff += np.linalg.norm(camNet.points3d[img_id, joint_id] - camNet.points3d[img_id, joint_id + 1]) - median_lengths[median_index]
-            new_diff += np.linalg.norm(points_using_2_cams[best_cam_tuple_index] - camNet.points3d[img_id, joint_id + 1]) - median_lengths[median_index]
-
-        if new_diff < old_diff:
-            #print("correcting", img_id, joint_id)
-            camNet.points3d[img_id, joint_id] = points_using_2_cams[best_cam_tuple_index]
-        #else:
-        #    pass
-            #print("not correcting", img_id, joint_id, new_diff, old_diff)
-        #if img_id == 18957:
-        #    print(segment_length_diff)
-        #    print(new_diff)
-        #    print(old_diff)
-        #    exit()
-    #print(np.histogram(np.max(camNet.points3d[:, :, 0], axis=1)))
-    #exit()
-    #center_frame_index = np.argmax(np.max(camNet.points3d[:, :, 0], axis=1))
-    #center_frame_index = 12801
-    #center_frame_index = 18957
-    #print(center_frame_index)
-    #camNet.points3d = camNet.points3d[max(0, center_frame_index - 50):min(camNet.points3d.shape[0], center_frame_index + 50)]
-    #exit()
-    #print(camNet.points3d.shape)
-    camNet.points3d = deepfly.signal_util.filter_batch(camNet.points3d, freq=100)
-    return camNet
-
 
 def load_data_to_dict(data, skeleton):
     final_dict ={}
